refactored code, and css
I refactored the code so that the PlayPage and the SandboxPlayPage can be created with components (as they had similar code). I added functionality for ability to easily switch between titles for the page. Fixed issue with CSS where it was not letting the doors align properly. Also fixed CSS issue with title where the title was not aligned properly.
